+++ libs/components/ag-grid/src/lib/modules/ag-grid/ag-grid-wrapper.component.spec.ts
@@ -42,7 +42,7 @@ describe('SkyAgGridWrapperComponent', () => {
   };
 
   const agGrid = {
-    api: new GridApi(),
+    api: jasmine.createSpyObj('api', GridApi.prototype),
     columnApi: new ColumnApi(),
     gridReady: new Subject<GridReadyEvent>(),
     rowDataUpdated: new Subject<RowDataUpdatedEvent>(),
@@ -121,9 +121,9 @@ describe('SkyAgGridWrapperComponent', () => {
   });
 
   it('should apply ag-theme', async () => {
-    spyOn(agGrid.api, 'setHeaderHeight').and.returnValue(undefined);
-    spyOn(agGrid.api, 'resetRowHeights').and.returnValue(undefined);
-    spyOn(agGrid.api, 'refreshCells').and.returnValue(undefined);
+    (agGrid.api.setHeaderHeight as jasmine.Spy).and.returnValue(undefined);
+    (agGrid.api.resetRowHeights as jasmine.Spy).and.returnValue(undefined);
+    (agGrid.api.refreshCells as jasmine.Spy).and.returnValue(undefined);
     expect(
       gridWrapperNativeElement.querySelector('.sky-ag-grid')
     ).toHaveCssClass('ag-theme-sky-data-grid-default');
@@ -220,7 +220,7 @@ describe('SkyAgGridWrapperComponent', () => {
     it('should not move focus when tab is pressed but cells are being edited', () => {
       const col = {} as Column;
       spyOn(gridAdapterService, 'setFocusedElementById');
-      spyOn(agGrid.api, 'getEditingCells').and.returnValue([
+      (agGrid.api.getEditingCells as jasmine.Spy).and.returnValue([
         { rowIndex: 0, column: col, rowPinned: undefined },
       ]);
 
@@ -232,8 +232,8 @@ describe('SkyAgGridWrapperComponent', () => {
     it('should not move focus when tab is pressed but master/detail cells are being edited', () => {
       const col = {} as Column;
       spyOn(gridAdapterService, 'setFocusedElementById');
-      spyOn(agGrid.api, 'getEditingCells').and.returnValue([]);
-      spyOn(agGrid.api, 'forEachDetailGridInfo').and.callFake((fn) => {
+      (agGrid.api.getEditingCells as jasmine.Spy).and.returnValue([]);
+      (agGrid.api.forEachDetailGridInfo as jasmine.Spy).and.callFake((fn) => {
         fn(
           {
             api: {
@@ -253,7 +253,7 @@ describe('SkyAgGridWrapperComponent', () => {
 
     it('should not move focus when a non-tab key is pressed', () => {
       spyOn(gridAdapterService, 'setFocusedElementById');
-      spyOn(agGrid.api, 'getEditingCells').and.returnValue([]);
+      (agGrid.api.getEditingCells as jasmine.Spy).and.returnValue([]);
 
       fireKeydownOnGrid('L', false);
 
@@ -262,8 +262,8 @@ describe('SkyAgGridWrapperComponent', () => {
 
     it(`should move focus to the anchor after the grid when tab is pressed, no cells are being edited,
       and the grid was previously focused`, () => {
-      spyOn(agGrid.api, 'getEditingCells').and.returnValue([]);
-      spyOn(agGrid.api, 'forEachDetailGridInfo').and.callFake((fn) => {
+      (agGrid.api.getEditingCells as jasmine.Spy).and.returnValue([]);
+      (agGrid.api.forEachDetailGridInfo as jasmine.Spy).and.callFake((fn) => {
         fn(
           {
             api: {
@@ -290,7 +290,7 @@ describe('SkyAgGridWrapperComponent', () => {
 
     it(`should move focus to the anchor before the grid when shift + tab is pressed, no cells are being edited,
       and the grid was previous focused`, () => {
-      spyOn(agGrid.api, 'getEditingCells').and.returnValue([]);
+      (agGrid.api.getEditingCells as jasmine.Spy).and.returnValue([]);
       spyOn(gridAdapterService, 'getFocusedElement').and.returnValue(
         skyAgGridDivEl
       );
@@ -323,7 +323,6 @@ describe('SkyAgGridWrapperComponent', () => {
     }
 
     it('should not move focus when tab is pressed but cells are being edited', () => {
-      spyOn(agGrid.api, 'stopEditing');
       fireKeyupEscape();
       expect(agGrid.api.stopEditing).toHaveBeenCalled();
     });
@@ -355,7 +354,6 @@ describe('SkyAgGridWrapperComponent', () => {
       spyOn(agGrid.columnApi, 'getAllDisplayedColumns').and.returnValue([
         column,
       ]);
-      spyOn(agGrid.api, 'ensureColumnVisible').and.stub();
 
       focusOnAnchor(afterAnchorEl, afterButtonEl);
 
@@ -374,8 +372,6 @@ describe('SkyAgGridWrapperComponent', () => {
       spyOn(agGrid.columnApi, 'getAllDisplayedColumns').and.returnValue([
         column,
       ]);
-      spyOn(agGrid.api, 'getFirstDisplayedRow');
-      spyOn(agGrid.api, 'setFocusedCell');
 
       focusOnAnchor(afterAnchorEl, afterButtonEl);