+++ libs/components/ag-grid/src/lib/modules/ag-grid/cell-editors/cell-editor-lookup/cell-editor-lookup.component.spec.ts
@@ -1,11 +1,23 @@
-import { ComponentFixture, TestBed } from '@angular/core/testing';
+import { ElementRef } from '@angular/core';
+import {
+  ComponentFixture,
+  TestBed,
+  fakeAsync,
+  tick,
+} from '@angular/core/testing';
 import { ReactiveFormsModule } from '@angular/forms';
 import { NoopAnimationsModule } from '@angular/platform-browser/animations';
 import { SkyAppTestUtility, expect } from '@skyux-sdk/testing';
 import { SkyInputBoxModule } from '@skyux/forms';
 import { SkyLookupModule, SkyLookupSelectModeType } from '@skyux/lookup';
 
-import { Column, ICellEditorParams, KeyCode } from 'ag-grid-community';
+import {
+  Column,
+  GridApi,
+  GridOptions,
+  ICellEditorParams,
+  KeyCode,
+} from 'ag-grid-community';
 import { of } from 'rxjs';
 
 import { SkyCellEditorLookupParams } from '../../types/cell-editor-lookup-params';
@@ -25,8 +37,12 @@ describe('SkyAgGridCellEditorLookupComponent', () => {
   let callback: ((args: Record<string, unknown>) => void) | undefined;
   const selection = [data[0]];
   let cellEditorParams: Partial<SkyCellEditorLookupParams>;
+  let elementRef: ElementRef<HTMLElement>;
 
   beforeEach(() => {
+    elementRef = {
+      nativeElement: jasmine.createSpyObj('nativeElement', ['matches']),
+    };
     TestBed.configureTestingModule({
       declarations: [SkyAgGridCellEditorLookupComponent],
       imports: [
@@ -35,9 +51,16 @@ describe('SkyAgGridCellEditorLookupComponent', () => {
         SkyInputBoxModule,
         SkyLookupModule,
       ],
+      providers: [
+        {
+          provide: ElementRef,
+          useValue: elementRef,
+        },
+      ],
     });
 
     cellEditorParams = {
+      api: jasmine.createSpyObj('api', GridApi.prototype),
       cellStartedEdit: true,
       colDef: {
         headerName: 'header',
@@ -52,7 +75,9 @@ describe('SkyAgGridCellEditorLookupComponent', () => {
           [event].pop();
         },
       } as Column,
-      context: undefined,
+      context: {
+        gridOptions: {} as Partial<GridOptions>,
+      },
       data: undefined,
       formatValue: jasmine.createSpy('formatValue'),
       onKeyDown: jasmine.createSpy('onKeyDown'),
@@ -209,7 +234,7 @@ describe('SkyAgGridCellEditorLookupComponent', () => {
           },
         });
 
-        expect(component.editorForm.get('selection')?.value).toBe(selection);
+        expect(component.editorForm.get('selection')?.value).toEqual(selection);
 
         fixture.detectChanges();
         await fixture.whenStable();
@@ -235,7 +260,7 @@ describe('SkyAgGridCellEditorLookupComponent', () => {
           eventKey: KeyCode.ENTER,
         });
 
-        expect(component.editorForm.get('selection')?.value).toBe(selection);
+        expect(component.editorForm.get('selection')?.value).toEqual(selection);
 
         fixture.detectChanges();
         await fixture.whenStable();
@@ -262,8 +287,29 @@ describe('SkyAgGridCellEditorLookupComponent', () => {
           charPress: 'a',
         });
 
-        expect(component.editorForm.get('selection')?.value).toBe(selection);
+        expect(component.editorForm.get('selection')?.value).toEqual(selection);
       });
+
+      it('should respond to focus changes', fakeAsync(() => {
+        cellEditorParams.context.gridOptions = {
+          stopEditingWhenCellsLoseFocus: true,
+        };
+        (elementRef.nativeElement.matches as jasmine.Spy).and.returnValue(
+          false
+        );
+        component.agInit({
+          ...(cellEditorParams as ICellEditorParams),
+        });
+
+        component.onLookupOpenChange(false);
+        tick();
+        expect(cellEditorParams.api?.stopEditing).toHaveBeenCalledTimes(1);
+
+        (cellEditorParams.api?.stopEditing as jasmine.Spy).calls.reset();
+        component.onBlur();
+        tick();
+        expect(cellEditorParams.api?.stopEditing).toHaveBeenCalledTimes(1);
+      }));
     });
 
     describe('cellStartedEdit is false', () => {
@@ -301,7 +347,7 @@ describe('SkyAgGridCellEditorLookupComponent', () => {
           eventKey: KeyCode.F2,
         });
 
-        expect(component.editorForm.get('selection')?.value).toBe(selection);
+        expect(component.editorForm.get('selection')?.value).toEqual(selection);
       });
 
       it('initializes with the current value when Enter triggers the edit', () => {
@@ -312,7 +358,7 @@ describe('SkyAgGridCellEditorLookupComponent', () => {
           eventKey: KeyCode.ENTER,
         });
 
-        expect(component.editorForm.get('selection')?.value).toBe(selection);
+        expect(component.editorForm.get('selection')?.value).toEqual(selection);
       });
 
       // NOTE: This is different than other editors due to the selection nature of autocomplete
@@ -324,7 +370,7 @@ describe('SkyAgGridCellEditorLookupComponent', () => {
           charPress: 'a',
         });
 
-        expect(component.editorForm.get('selection')?.value).toBe(selection);
+        expect(component.editorForm.get('selection')?.value).toEqual(selection);
       });
     });
   });