+++ libs/components/ag-grid/src/lib/modules/ag-grid/cell-editors/cell-editor-lookup/cell-editor-lookup.component.ts
@@ -4,6 +4,7 @@ import {
   Component,
   ElementRef,
   HostBinding,
+  HostListener,
 } from '@angular/core';
 import { UntypedFormControl, UntypedFormGroup } from '@angular/forms';
 
@@ -44,6 +45,7 @@ export class SkyAgGridCellEditorLookupComponent
 
   protected ariaLabel: string | undefined = undefined;
 
+  #lookupOpen = false;
   #params: SkyCellEditorLookupParams | undefined;
   #triggerType: SkyAgGridCellEditorInitialAction | undefined;
   #changeDetector: ChangeDetectorRef;
@@ -52,6 +54,11 @@ export class SkyAgGridCellEditorLookupComponent
   constructor(changeDetector: ChangeDetectorRef, elementRef: ElementRef) {
     this.#changeDetector = changeDetector;
     this.#elementRef = elementRef;
+  }
+
+  @HostListener('blur')
+  public onBlur(): void {
+    this.#stopEditingOnBlur();
   }
 
   public agInit(params: SkyCellEditorLookupParams): void {
@@ -142,10 +149,25 @@ export class SkyAgGridCellEditorLookupComponent
     }
   }
 
+  public onLookupOpenChange(isOpen: boolean): void {
+    this.#lookupOpen = isOpen;
+    this.#stopEditingOnBlur();
+  }
+
   #updateComponentProperties(
     params: SkyCellEditorLookupParams
   ): SkyAgGridLookupProperties {
     const skyLookupProperties = params.skyComponentProperties;
     return applySkyLookupPropertiesDefaults(skyLookupProperties);
+  }
+
+  #stopEditingOnBlur(): void {
+    if (
+      !this.#lookupOpen &&
+      this.#params?.context?.gridOptions?.stopEditingWhenCellsLoseFocus &&
+      !this.#elementRef.nativeElement.matches(':focus-within')
+    ) {
+      this.#params?.api.stopEditing();
+    }
   }
 }