+++ libs/components/lookup/src/lib/modules/autocomplete/autocomplete.component.ts
@@ -315,6 +315,12 @@ export class SkyAutocompleteComponent implements OnDestroy, AfterViewInit {
   @Output()
   public searchAsync = new EventEmitter<SkyAutocompleteSearchAsyncArgs>();
 
+  /**
+   * @internal
+   */
+  @Output()
+  public openChange = new EventEmitter<boolean>();
+
   //#endregion
 
   //#region template_properties
@@ -553,6 +559,7 @@ export class SkyAutocompleteComponent implements OnDestroy, AfterViewInit {
     this.#ngUnsubscribe.complete();
     this.#destroyAffixer();
     this.#destroyOverlay();
+    this.openChange.complete();
   }
 
   public addButtonClicked(): void {
@@ -883,6 +890,7 @@ export class SkyAutocompleteComponent implements OnDestroy, AfterViewInit {
       this.#changeDetector.markForCheck();
       this.#updateAriaControls();
       this.#initOverlayFocusableElements();
+      this.openChange.emit(true);
     }
   }
 
@@ -893,6 +901,7 @@ export class SkyAutocompleteComponent implements OnDestroy, AfterViewInit {
     this.#removeActiveDescendant();
     this.#updateAriaControls();
     this.#changeDetector.markForCheck();
+    this.openChange.emit(false);
   }
 
   #setActiveDescendant(): void {